watchlist app
I usually spent a lot of time searching what to watch when I want to watch a movie. So, I thought it would be pretty convenient for me todesign a website that will help me hasten my decison making when choosing a movie. This app is specialized for me and currently it only has 2 movies (Inception and Godfather).I will try my best to expand the list with movies that I haven't watch.